test: run shell secret tests in CI
Consolidate shell secret coverage into one cross-platform test file and run it explicitly in the CI platform matrix.
