+++ .github/workflows/ci.yml
@@ -168,6 +168,9 @@ jobs:
         if: runner.os == 'Linux'
         uses: t1m0thyj/unlock-keyring@v1
 
+      - name: Run shell secret tests
+        run: bun test src/tests/tools/shell-secrets.test.ts --timeout 15000
+
       - name: Run tests (extended timeout)
         # Unit tests must pass for fork PRs (no secrets). Keep API-dependent tests
         # in a separate gated step.
